up dates deps and maybe some other things?

+++ structure/Settings/LegalitySettings.cs
@@ -17,9 +17,9 @@ namespace _3DS_link_trade_bot
         [Category(Legality), Description("Bots OT name, must be less than 8 characters for past generations to work")]
         public string BotOT { get; set; } = "pip";
         [Category(Legality), Description("Bots 5 digit TID, Not the 6 digit Trainer ID 7")]
-        public int BotTID { get; set; } = 42069;
+        public ushort BotTID { get; set; } = 42069;
         [Category(Legality), Description("Bots 5 digit SID, Not the 4 digit SID7")]
-        public int BotSID { get; set; } = 42069;
+        public ushort BotSID { get; set; } = 42069;
         [Category(Legality), Description("Bots Game Language")]
         public PKHeX.Core.LanguageID BotLanguage = PKHeX.Core.LanguageID.English;
         [Category(Legality), Description("add all Legal Ribbons to showdown set trades")]
